+++ relay-server/src/actors/processor.rs
@@ -2369,18 +2369,6 @@ impl EnvelopeProcessorService {
                     .map(|(k, v)| (k.sentry_tag_key().to_owned(), Annotated::new(v)))
                     .collect(),
             );
-            // Double write to `span.data` for now. This can be removed once all users of these fields
-            // have switched to `sentry_tags`.
-            let data = transaction_span
-                .data
-                .value_mut()
-                .get_or_insert_with(Default::default);
-            data.extend(
-                shared_tags
-                    .clone()
-                    .into_iter()
-                    .map(|(k, v)| (k.data_key().to_owned(), Annotated::new(v.into()))),
-            );
             add_span(transaction_span.into());
         }
     }